+++ views/admin/index.php
@@ -202,6 +202,11 @@ use yii\web\View;
         </div>
 
         <div class="form-group">
+            <?= Html::activeLabel($model, 'commonLabel', ['class' => 'control-label']); ?>
+            <?= $form->field($model, 'openInNewTab')->checkbox(); ?>
+        </div>
+
+        <div class="form-group">
             <?= Html::Button(
                 Yii::t(
                     'OnlyofficeModule.base',
@@ -287,6 +292,7 @@ use yii\web\View;
                     var type = $(node).attr("id").replace("configureform-forceedittypes-", "");
                     forceEditTypes[type] = $(node).prop("checked") ? 1 : 0;
                 });
+                var openInNewTab = $("#configureform-openinnewtab").prop("checked") ? 1 : 0;
 
                 $.ajax({
                     url: "' . Url::to(["/onlyoffice/admin/save"]) . '",
@@ -307,7 +313,8 @@ use yii\web\View;
                             feedback: feedback,
                             help: help,
                             compactToolbar: compactToolbar,
-                            forceEditTypes: forceEditTypes
+                            forceEditTypes: forceEditTypes,
+                            openInNewTab
                         }
                     },
                     dataType: "json"