+++ widgets/views/editor.php
@@ -28,8 +28,11 @@ if ($documentType === Module::DOCUMENT_TYPE_SPREADSHEET) {
 ?>
 
 <?= Html::beginTag('div', $options) ?>
-<div style="height:50px;
-    border-radius: 5px 5px 0px 0px;
+<div style="<?php if ($openInNewTab) :
+    ?>display:none;<?php
+            endif; ?>
+    height:50px;
+    border-radius:5px 5px 0px 0px;
     background-color:<?= $headerBackgroundColor; ?>;
     padding-top:7px;
     padding-right:7px">
@@ -38,7 +41,9 @@ if ($documentType === Module::DOCUMENT_TYPE_SPREADSHEET) {
             <?= humhub\libs\Html::a(
                 Yii::t('OnlyofficeModule.base', 'Share'),
                 '#',
-                ['class' => 'btn btn btn-default',
+                [
+                    'id' => 'onlyoffice-share-link',
+                    'class' => 'btn btn btn-default',
                     'data-action-click' => 'share',
                     'data-action-block' => 'sync',
                     'data-action-url' => Url::to([