+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ##
 ## Added
+- openInNewTab admin setting to open the editor in a new tab instead of a modal
+- opening file location from editor in tab
 - plugin description and useful links in admin settings
 
 ## Changed